Removes need for `g:ncm_r_dt` option
`data.table` objects are detected automatically using the omni object
match list.

Also:

- Adds support for completion when using `data.table`s with `magrittr`
pipes, using `.` to refer to the 'parent' DT (note that as a
side-effect, ordinary `data.frame`s can be used in the same
way and completion work as expected)

- In order to get this pipe behaviour to work, the regular expression
for checking if completion is happening inside the brackets of a
`data.frame` was changed to allow dots in the name of the DF. This could
be considered a bug fix since previously completion inside brackets did
not work as expected if the DF had dots in its name.